Refactor: move EnergyOrb to `src` module, add `image` param; add type hints
